SDL library rework.
Found problems with the linux build, I guess the standard SDL lib installs on Linux are in (for example) #include <SDL2/SDL.h>. Specifically under the SDL2 include folder.

Changed the folder structure of the SDL libs to have an extra SDL2 folder to match the linux include names. Not tested on Windows yet.