SDL library include fixup.
In linux, added the SDL libraries to the lib folder even though we use the system libraries. This is because the system libraries are loaded in the <SDL/SDL.h> SDL/ subdirectory. This is not like in the Windows version that isn't in the SDL subdirectory. SDL_Image and SDL_ttf use the version that is not in the subdirectory so to compensate we have to switch back to the old version.