Geometry class update.
The geometry code was moved to the Libtech project. Pointed the
references to that.

Also added the same fix as in Libtech to set the DLLexport property on
ENGINE_CLASS in case consumers of that class were using non-pointer
types of those classes.

Small change in CMakeLists to have the MicroEngine DLL link to the
Libtech DLL library.