Adding new game project + refactor.
Added the RPG-Game project, this project will help the engine advance
by prototyping a game and adding engine functionality as we go.

Fixed a bug in the window title that was caused by the c_str() of the
config search to be overridden by last const char taking the place
of the window title address. It seems new std::string built on the stack
will have their c_str() occupy the same address. Also setting the
default's window title be not-const because it might change.