Texture pipeline upgrade.
Textures can now set themselves up by loading an image or setting a
solid color. This is allowed because they keep a reference to the
renderer.

Its not optimal because the new SDL_Surface and SDL_Texture instances
are not tracked by the renderer so the texture effectively owns the
memory. Bad for tracking but ok for now.

Added CreateTexture functions to GameEngine and GameModule to avoid
having to all the way down to the texture repository.