Texture Repository rework
Completely changed the way TextureRepository works. It is no longer
just a front for a FileCache instance and is not in charge of creating
instances of texture.

The new role of TextureRepository is just to hold instances of ATexture
and allow us to check if a texture is already loaded before creating a
new instance of it. Since it is no longer involved in the specific
details of creating textures it is not subclassed by SDL anymore and
gone from the Abstract Graphic Layer.

The Renderer owns the TextureRepository and looks inside when a caller
asks for a texture. This way, no one has to know anything about the
texture repository when creating textures.

The file cache that was in the repo is now created by the GameEngine
and pointers are given to classes that needs it. That way everyone
shares the same file cache data.

Textures will soon be able to delete their data from memory while
leaving it uploaded on the GPU. Can be used to free up the program heap
under memory pressure. Not entirely implemented yet.