Added Viewport management.
Viewports are used to view sprites that are farther than the window's
view area. Viewports are like a camera over the world, when the camera
moves it looks at objects that are farther away.

The GameEngine has one selected viewport, it can be switched any time.
When rendering, it will take into account the currently selected
viewport.

With this update, Sprites have a variable called a Positioning System
(the name might change). By default it is set to VIEWPORT_RELATIVE,
that means it will move relative to the where the viewport is.
Sometimes we want to stick a sprite to a particular position on the
window and not move with the viewport like the UI or some title header.
For that we can set the sprite's positioning system to VIEWPORT_STATIC.

I had to add a new method to draw a texture. This is because the
positioning system variable is set at the sprite level, this would make
single texture draws unable to draw relative to the viewport. A texture
draw option can be used to indicate that the texture is relative to
wherever the viewport is at.