Warnings fixes
Fixed the strings warning in DLL exports by transforming the text field of an AText instance to a pointer. Not sure what the warning was about but it was probably because of an export of a string using a particular C runtime over to another. Now it uses a pointer managed by the instance.

Fixed compiler warnings on linux. Its a wonder things compile at all on Windows...