Debug layer integration.
The debug layer is the information that can be drawn on top of the game
window. The debug layer can be activated by calling
engine->ShowDebugLayer. Typically only the engine handles the
information of the debug layer but the game will be able to add values
in there.

Several engine improvements also :
- Moved the size and position methods to the base class of the
renderable objects.

- Added methods to release object memory. I'm thinking to use the
ReleaseObject methods to delete objects from the Abstract Graphic Layer
since the memory is going to be managed differently, releasing a
resource is going to more involved than the destructor of an object.

- Added Enable and Disable methods to a GameModule so it gets called
when the game goes to and from a module.

- Fixed the IsKeyPressed method from the keyboard class.

- Added a check to stop drawing a Sprite if it is not visible.

- Fixed the SetSolidColor function on textures that needed to have the
color masks to work. Also switched to SDL_FreeSurface because simply
calling SDL_free does not work.