Capped update events to once per frame.
This change was forced because of the input loop was out of sync with
the classes polling. This is due to the input loop being able to roll
multiple times before the game classes were updated once per 60hz. That
was making press/release events pass before the game could poll those
changes. Now press and release can happen at max once per frame.

The concern with that method is that events that happen in sub 60hz
frequency may be missed because we don't poll as much. We'll see if
this is a problem eventually.

This commit enables the debug view to be opened by pressing F1.