Work on XFile functionality.
Using Win32 for the file IO abstraction because the IO classes are
meant to be passed around and used by many consumers. Specifically I'm
not using C++ standard IO with FILE* or with streams because I want to
be able to customize the IO processes by OS. Also, because its a great
way to learn about win32.