Started adding an Asset class.
This new class will allow loading files in a more dynamic manner.
Currently, loading assets is done by sending a path to some function
that will do the loading. With an Asset instance, an asset can be
passed to objects without having the data loaded.  An asset can be
typed and can have advanced features such as being memory mapped.

Work in progress and not tested yet, will eventually plug into the
ResourceManager and objects such as an AFont will have an overload to
work with Asset types.