Input Controls workable prototype.
Added logic for the button class and base control rework.

Controls now inherit from ARenderable because all objects in the engine that are meant to be drawn must inherit from that class.

The SDL Renderer was reworked to stop converting the renderable parameter to a sprite, that was unnecessary and stupid.

Added some test images for a button for playing around.