Added Button click testing
When the button is down and enabled, calling IsClicked will return true.

I added an optional parameter in the IsClicked method to avoid repeating the result on each call to the function. If we check with a true parameter, the click result will only be true once. Releasing the button is required to get another true result.