Sprite Animations basic implementation.
Many changes were needed to support the implementation.

- New method of Textures to create a sub texture from source pixels in
a texture.

- Added method in the Renderer to delete a texture. This is to allow a
hook to allow to deallocate managing resources of a texture not
directly removed when calling delete(texture).

- Added the FrameListLoader class. This class's job is to take a path
as input and produce a list of output used for loading a sprite's frame
animations. Can return a list of files path, a list of textures, a list
of animations. The input path can be a directory to load all the images
within and make an animation or read in a spritesheet and create
individual textures from it. In the future a folder can have a
"manifest" file to guide the loading of each frame inside the folder.

- Added a clock animation to test the loading and playback mechanism.

Many functions are not implemented yet but will be expanded upon as
needed.