Fixed FPS bug.
The bug was caused by the storage type of a Sprite’s previousFrameTime value. The millisecond values returned by the get_a_ticks functions is unsigned. This is in line with the type returned by the POSIX functions. Not sure why they return an unsigned value but there we go.