Sprite animation API pass
Work on the ease of access to play animations. Added a dinosaur spritesheet to test frames creation and handling. More information follows lower.

Added a GameObject base class. This class serves for game classes to provide them with the standard Setup/Update/Draw functions. Implementing the class is not required and the engine doesn’t care about inheritance of that class but it lets us know that these 3 methods are going to be the main callsite from the module into the game classes.

SpriteAnimation was switched to a class to allow greater logic to be done from inside the class. Added a function to set the name to avoid having to create a new char* from the caller and to strcpy the wanted name into the variable. Also added static functions to create a sprite animation from a list of texture and a string pattern like “0..12” to take the first 12 frames of the list. This allows us to have a list of animations and to select parts of them for a single sprite animation.

Changed the parameters for the Sprite Play function for something that makes more sense. Calling it with a NULL animation name will play the first animation with no name.

Added functions in the Texture class to load a spritesheet with the images side-by-side and split it by numerous ways.

Added a wrapper function in GameModule to destroy objects once we’re done with it.

Trailing whitespace was cleaned off some files that I was working on because they were saved from Xcode.

Added a DinoCharacter class in the engine tester to test the new animations/frames API.