Added ARenderable scaling
Also messed with texture management. Since we set the destination
rectangle using the size of the sprite instead of the size of the
texture, some sprites were broken in the tester because they were set
from images and setting a texture from an image does not update the
source sprite size. Now setting an animation or a image texture on a
ARenderable will update the size.

Cleaned some places where the * of a pointer was not touching the type
name.