Managed to specialize the Event class for void
Turns out it wasn’t as hard as I thought. Two instances of both classes have been created, this creates a bit of code duplication but since I need to adapt the inner std::function member of the class I think I need to create a whole copy of the class.

Both classe are the same except they deal with passing 1 or no argument to the dispatch function and declared the std::function with a parameter type or not.