Textures rework
Changed the way SetSize() of a texture works. Now, the texture will execute a resize and scale the contents up or down.

This is in preparation of being able to copy part of a texture into part of another texture for making a texture strip from a list of textures.

The SDLTexture had an extra ‘TextureSize’ that was changed at some places. It was removed because the parent ATexture class is carrying the only reference to texture size.