Added screenshotting
The two new functions of the Renderer can be used to save the next frame to an image on disk and to save a specific renderable object.

Can be useful to debug the whole scene because they draw the entire texture and not just the part shown in the window.