Added Set Texture overload on Renderable
The overload to create a sprite from a string that can take the form of an asset name of file path was only available on ASprite.

For some reason, I can't remove the overload from ASprite and keep the same one in RenderableObject as some callers can't resolve the proper function ? No idea why.