Started to introduce GameModule loading
Added states to GameModules in order to be able to load them and
activate them. That will enable them to receive updates and draw
calls.

Still working out the API of the modules and how they can
communicate with one another.

Fixed a call to RenderableObject::SetTexture where it wasn't
setting the size of an object when setting a texture on it.

Removed a spammy LogTrace call when doing a texture FillColor
on an SDL Texture.