MacOS SDL library fix
Due to security on MacOS requiring that packages and frameworks be
signed SDL_image was unable top be loaded. Rather than waiting for
the dev to sign his package, I changed the way that MacOS loads
libraries.

Instead of using a Mac Framework for SDL, I chose to switch over to
dynamic libraries downloaded through homebrew. This is far simpler
than dealing Frameworks and don't require the libraries to be signed.