Renderable children draw by default
Now activated automatically drawing an object's children. Previously
it was a goal to separate the engine rendering mechanism from the
parent->child relationship. Maybe even to support another rendering
scheme.

After playing with the engine a while, using Parent as containers for
children seem to be the easier approach.

Removed the Module's child rendering system since it is now taken
into account by the renderer.